+++ src/main/java/com/nirmaan/project/entity/KycModel.java
@@ -18,7 +18,7 @@ public class KycModel {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int kyc_Id;
+	private long kyc_Id;
 
 	@ManyToOne
 	@JsonBackReference
